Locator changes. More functionality (deduplication).
